remove the cast to (int) - column_width is a double.
2002-11-18  Chris Toshok  <toshok@ximian.com>

	* gui/widgets/gal-view-minicard.c (gal_view_minicard_attach):
	remove the cast to (int) - column_width is a double.

	* gui/widgets/e-minicard.c (e_minicard_class_init): the height
	property is only readable, not readwrite.

	* gui/widgets/e-minicard-view-widget.c
	(e_minicard_view_widget_class_init): make column_width a double
	property.
	(e_minicard_view_widget_set_property): don't cast to (int).
	column_width is double in the e-minicard-view.
	(e_minicard_view_widget_realize): same.
	(e_minicard_view_widget_get_property): use g_value_set_double for
	column_width.

svn path=/trunk/; revision=18831

+++ addressbook/gui/widgets/gal-view-minicard.c
@@ -197,7 +197,7 @@ gal_view_minicard_attach (GalViewMinicard *view, EMinicardViewWidget *emvw)
 	g_object_ref (view->emvw);
 
 	g_object_set (view->emvw,
-		      "column_width", (int) view->column_width,
+		      "column_width", view->column_width,
 		      NULL);
 
 	view->emvw_column_width_changed_id =
